refactor: 폴더 구조 번경 및 예시 폴더 생성
