+++ package.json
@@ -31,6 +31,7 @@
     "stomp": "^0.1.1"
   },
   "devDependencies": {
+    "@types/node": "^22.13.5",
     "@types/react": "^18.2.43",
     "@types/react-dom": "^18.2.17",
     "@types/react-router-dom": "^5.3.3",
