+++ src/components/chat/chat-room/Chat.tsx
@@ -1,179 +1,36 @@
 import React, { useEffect, useRef, useState } from 'react'
-import styled from 'styled-components'
 // import attachImg from '../assets/images/attach.png'
 // import photoImg from '../assets/images/photo.png'
-import Stomp, { Client } from '@stomp/stompjs'
+import {Client} from '@stomp/stompjs'
 import axios from 'axios'
 import profileImg from '../../assets/images/profileimg.png'
-import { useApiUrlStore } from '../../store/store'
+import { useApiUrlStore } from "store/store"
+import { ChatProps, ContentProps} from "./Chat.type"
+import {
+  Container,
+  ChatWrap,
+  MessageContainer,
+  Profile,
+  Messages,
+  InputWrap,
+  InputField,
+  SendButton,
+  BtnWrap,
+  ZoomLoginBtn,
+  CreateMeetingBtn,
+  CreateReviewBtn,
+} from "./Chat.styles" 
 
-interface ChatProps {
-  chatRoomId: string
-  onOpen: () => void
-}
 
-interface Content {
-  type: string
-  chatRoomId: string
-  sender: string
-  content: string
-}
-
-interface MessagesProps {
-  sender?: string
-  nickname: string
-}
-
-interface ProfileProps {
-  sender?: string
-  nickname: string
-  src: string
-}
-
-interface MessageContainerProps {
-  sender?: string
-  nickname: string
-}
-
-const Container = styled.div`
-  width: 95.6rem;
-  height: calc(100vh - 17.5rem);
-  margin-top: 1.25rem;
-  margin-bottom: 1.25rem;
-  border-left: 2px solid #d8d8d8;
-  border-right: 2px solid #d8d8d8;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-`
-
-const ChatWrap = styled.div`
-  width: 95.6rem;
-  height: calc(100vh - 12.5rem);
-  overflow-y: auto; /* 내용이 넘칠 때만 스크롤 표시 */
-`
-
-const MessageContainer = styled.div<MessageContainerProps>`
-  display: flex;
-  align-items: center;
-  justify-content: ${(props) => (props.sender === props.nickname ? 'flex-end' : 'flex-start')};
-  margin-top: 0.75rem;
-  margin-bottom: 0.75rem;
-`
-
-// const Time = styled.span`
-//   font-size: 12px;
-//   color: #888; /* Gray color for time */
-// `
-
-const Profile = styled.img<ProfileProps>`
-  padding: 0.8rem;
-  width: 6.25rem;
-  height: 6.25rem;
-  margin-left: ${(props) => (props.sender !== props.nickname ? '0' : '0.625rem')};
-  margin-right: ${(props) => (props.sender !== props.nickname ? '0.625rem' : '0')};
-`
-
-const Messages = styled.div<MessagesProps>`
-  display: flex;
-  max-width: 40%;
-  padding: 1.25rem;
-  margin-right: 0.625rem;
-  height: auto;
-  justify-content: center;
-  align-items: center;
-  background: ${(props) =>
-    props.sender !== props.nickname ? 'rgba(231, 227, 227, 0.8)' : '#8a33cb'};
-  color: ${(props) => (props.sender !== props.nickname ? 'black' : 'white')};
-  border-radius: 1rem;
-  font-size: 1.6rem;
-  margin-left: ${(props) => (props.sender !== props.nickname ? '0.625rem' : 'auto')};
-`
-
-const InputWrap = styled.div`
-  justify-content: center;
-  width: 100%;
-  display: flex;
-  align-items: center;
-`
-
-const InputField = styled.input`
-  height: 4.375rem;
-  width: 75rem;
-  border-radius: 3rem;
-  border: none;
-  font-size: 1.25rem;
-  background-color: #f6f6f6;
-  padding-left: 1.8rem;
-  font-size: x-large;
-`
-
-const SendButton = styled.button`
-  width: 7.5rem;
-  height: 4.375rem;
-  margin-left: 3rem;
-  border-radius: 1.25rem;
-  font-size: x-large;
-  font-weight: bold;
-  border: none;
-  background-color: #650fa9;
-  color: #ffffff;
-  cursor: pointer;
-`
-const BtnWrap = styled.div`
-  display: flex;
-  width: 100%;
-  justify-content: right;
-  border-bottom: 2px solid #d8d8d8;
-`
-
-const ZoomLoginBtn = styled.button`
-  width: 7.5rem;
-  height: 4.375rem;
-  border-radius: 1.25rem;
-  font-size: large;
-  font-weight: bold;
-  border: none;
-  background-color: #650fa9;
-  color: #ffffff;
-  cursor: pointer;
-  margin: 0rem 0.625rem 1.25rem 0.625rem;
-`
-const CreateMeetingBtn = styled.button`
-  width: 7.5rem;
-  height: 4.375rem;
-  border-radius: 1.25rem;
-  font-size: x-large;
-  font-weight: bold;
-  border: none;
-  background-color: #650fa9;
-  color: #ffffff;
-  cursor: pointer;
-  margin: 0rem 0.625rem 1.25rem 0.625rem;
-`
-
-const CreateReviewBtn = styled.button`
-  width: 7.5rem;
-  height: 4.375rem;
-  border-radius: 1.25rem;
-  font-size: x-large;
-  font-weight: bold;
-  border: none;
-  background-color: #e8dcf2;
-  color: #650fa9;
-  cursor: pointer;
-  margin: 0rem 0.625rem 1.25rem 0.625rem;
-`
 
 function Chat({ chatRoomId, onOpen }: ChatProps) {
   const { apiUrl } = useApiUrlStore()
 
   const [nickname, setNickname] = useState('')
 
-  const [stompClient, setStompClient] = useState<Stomp.Client | null>(null)
+  const [stompClient, setStompClient] = useState<Client | null>(null)
 
-  const [messages, setMessages] = useState<Content[]>([])
+  const [messages, setMessages] = useState<ContentProps[]>([])
   const [inputMessage, setInputMessage] = useState('')
 
   const chatRef = useRef<HTMLDivElement>(null)
@@ -261,7 +118,7 @@ function Chat({ chatRoomId, onOpen }: ChatProps) {
 
   const sendMessage = (messageContent: string, nickname: string, messageType: string) => {
     const destination = `/pub/chat.message.${chatRoomId}`
-    const newMessage: Content = {
+    const newMessage: ContentProps = {
       type: messageType,
       chatRoomId: chatRoomId,
       sender: nickname,
