+++ src/components/study/components/StudyPostingModal.tsx
