+++ src/pages/mypage/UpdateProfilePage.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components'
 import defaultImg from '../../assets/images/profileimg.png'
 import Header from '../../components/Header.tsx'
 import Navbar from '../../components/Navbar.tsx'
-import Profilebar from '../../components/sidebar/Profilebar.tsx'
+import Profilebar from '../../shared/components/sidebar/Profilebar.tsx'
 import { getImageImageUrl, useApiUrlStore, useProfileDataStore } from '../../store/store.ts'
 
 const Container = styled.div`
