+++ src/shared/components/skeleton/DetailSkeletonUI.tsx
