+++ tsconfig.json
@@ -18,7 +18,9 @@
     "strict": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
-    "noFallthroughCasesInSwitch": true
+    "noFallthroughCasesInSwitch": true,
+
+    "baseUrl": "src",
   },
   "include": ["src"],
   "references": [{ "path": "./tsconfig.node.json" }]
